requirements.txt
laspy[lazrs]
open3d
pillow
plyfile
scipy
shapely
tm35fin
trimesh
tqdm
